Update to latest core changes
Core update 2.0

Issue: PLCORE20-16
