+++ src/controllers/admin/OrderDraftController.php
@@ -1,11 +1,7 @@
 <?php
 
 use Logeecom\Infrastructure\ServiceRegister;
-use Logeecom\Infrastructure\TaskExecution\QueueService;
-use Packlink\BusinessLogic\Order\Interfaces\OrderRepository as OrderRepositoryInterface;
-use Packlink\BusinessLogic\Order\Models\OrderShipmentDetails;
-use Packlink\BusinessLogic\Tasks\SendDraftTask;
-use Packlink\PrestaShop\Classes\Repositories\OrderRepository;
+use Packlink\BusinessLogic\ShipmentDraft\ShipmentDraftService;
 use Packlink\PrestaShop\Classes\Utility\PacklinkPrestaShopUtility;
 
 /** @noinspection PhpIncludeInspection */
@@ -18,40 +14,17 @@ class OrderDraftController extends PacklinkBaseController
 {
     /**
      * Creates order draft for order identified by ID in the request by enqueuing SendDraftTask.
-     *
-     * @throws \Logeecom\Infrastructure\ORM\Exceptions\QueryFilterInvalidParamException
-     * @throws \Logeecom\Infrastructure\ORM\Exceptions\RepositoryNotRegisteredException
-     * @throws \PrestaShopDatabaseException
-     * @throws \PrestaShopException
      */
     public function displayAjaxCreateOrderDraft()
     {
         $data = PacklinkPrestaShopUtility::getPacklinkPostData();
 
         if ($data['orderId']) {
-            $orderId = $data['orderId'];
-            /** @var QueueService $queue */
-            $queue = ServiceRegister::getService(QueueService::CLASS_NAME);
-            /** @var OrderRepository $orderRepository */
-            $orderRepository = ServiceRegister::getService(OrderRepositoryInterface::CLASS_NAME);
-
-            $orderDetails = $orderRepository->getOrderDetailsById($orderId);
-            if ($orderDetails === null) {
-                $orderDetails = new OrderShipmentDetails();
-                $orderDetails->setOrderId($orderId);
-                $orderRepository->saveOrderDetails($orderDetails);
-            }
-
             try {
-                $draftTask = new SendDraftTask($orderId);
-
-                $queue->enqueue($this->getConfigService()->getDefaultQueueName(), $draftTask);
-
-                if ($draftTask->getExecutionId() !== null) {
-                    $orderDetails->setTaskId($draftTask->getExecutionId());
-                    $orderRepository->saveOrderDetails($orderDetails);
-                }
-            } catch (\Logeecom\Infrastructure\TaskExecution\Exceptions\QueueStorageUnavailableException $e) {
+                /** @var ShipmentDraftService $shipmentDraftService */
+                $shipmentDraftService = ServiceRegister::getService(ShipmentDraftService::CLASS_NAME);
+                $shipmentDraftService->enqueueCreateShipmentDraftTask($data['orderId']);
+            } catch (\Exception $e) {
                 PacklinkPrestaShopUtility::die500(array(
                     'success' => false,
                     'message' => $e->getMessage(),
